examples/addsvc2: rename pkg/xxx to pkg/addxxx
