+++ examples/addsvc2/pkg/addendpoint/middleware.go
@@ -1,4 +1,4 @@
-package endpoint
+package addendpoint
 
 import (
 	"context"
