+++ examples/addsvc2/pkg/addservice/middleware.go
@@ -1,4 +1,4 @@
-package service
+package addservice
 
 import (
 	"context"
