+++ examples/addsvc2/pkg/addservice/service.go
@@ -1,4 +1,4 @@
-package service
+package addservice
 
 import (
 	"context"
